Update streamlit_app.py
Since we open a Snowflake Connection each time a button on the page is clicked, let's also close the connection at the end of each button's script (line 58, below). We've illustrated the first one so that you ca use it as an example and add a similar statement to the second button script.